fast_api/application/sms_api/requirements.txt
bcrypt==3.2.0
fastapi==0.109.0
parameterized==0.9.0
python-dotenv==1.0.0
uvicorn==0.25.0
SQLAlchemy==2.0.23
mysql-connector-python==8.0.33
